refactor: menu items
