+++ src/components/MenuBuilder/MenuBuilder.tsx
@@ -1,26 +1,29 @@
-import {
-  FormProvider,
-  SubmitHandler,
-  useFieldArray,
-  useForm,
-} from 'react-hook-form';
+import { FormProvider, useForm, SubmitHandler } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
+import { MenuFormInputs, menuFormSchema } from '@/schemas/menuForm';
+
 import { useSetAtom } from 'jotai';
-
-import MenuHeader from './MenuHeader/MenuHeader';
-import MenuItem from './MenuItem/MenuItem';
-
 import { menuAtom } from '@/store/menu.atom';
 
-import { MenuFormInputs, menuFormSchema } from '@/schemas/menuForm';
+import MenuHeader from './MenuHeader/MenuHeader';
+import MenuItems from './MenuItems/MenuItems';
+
 import { Button, Card } from 'react-daisyui';
-import { FiPlus } from 'react-icons/fi';
 
 import { useModal } from '@/hooks/useModal';
 import MenuModal from './MenuModal/MenuModal';
 
 const MenuBuilder = () => {
   const setMenuAtom = useSetAtom(menuAtom);
+
+  const methods = useForm<MenuFormInputs>({
+    resolver: yupResolver(menuFormSchema),
+    defaultValues: {
+      products: [{}],
+    },
+  });
+
+  const { control, handleSubmit, reset } = methods;
 
   const {
     modal: menuModal,
@@ -30,35 +33,13 @@ const MenuBuilder = () => {
     children: <MenuModal closeModal={() => closeModal()} />,
   });
 
-  const methods = useForm<MenuFormInputs>({
-    resolver: yupResolver(menuFormSchema),
-    defaultValues: {
-      products: [{}],
-    },
-  });
-
-  const { handleSubmit, reset, control } = methods;
-
-  const { fields, append, move, remove } = useFieldArray({
-    control,
-    name: 'products',
-  });
-
-  const addNewItem = () => {
-    append({
-      name: '',
-      description: '',
-      price: 0,
-    });
-  };
-
-  const resetForm = () => {
-    reset();
-  };
-
   const onFormSubmit: SubmitHandler<MenuFormInputs> = (values) => {
     setMenuAtom(values);
     openModal();
+  };
+
+  const onFormReset = () => {
+    reset();
   };
 
   return (
@@ -73,31 +54,9 @@ const MenuBuilder = () => {
               onSubmit={handleSubmit(onFormSubmit)}>
               <div className="flex flex-col gap-8 w-full">
                 <MenuHeader />
-                <div className="flex flex-col gap-2">
-                  <div className="flex justify-end px-2">
-                    <Button
-                      type="button"
-                      size="sm"
-                      shape="square"
-                      color="primary"
-                      onClick={addNewItem}>
-                      <FiPlus size={20} />
-                    </Button>
-                  </div>
-                  {fields.map((field, index) => {
-                    return (
-                      <MenuItem
-                        key={field.id}
-                        itemIndex={index}
-                        itemAmount={fields.length}
-                        removeItem={remove}
-                        moveItem={move}
-                      />
-                    );
-                  })}
-                </div>
+                <MenuItems control={control} arrayName="products" />
                 <div className="card-actions justify-end">
-                  <Button type="button" color="neutral" onClick={resetForm}>
+                  <Button type="button" color="neutral" onClick={onFormReset}>
                     Reset
                   </Button>
                   <Button type="submit" color="primary">
